refactor: useInputShell의 3가지 책임을 각 커스텀훅으로 분리
