+++ my-app/src/components/common/commonHooks/useFocusRule.ts
@@ -1,0 +1,13 @@
+import { useRef } from 'react';
+
+export const useFocusRule = (totalLength: number) => {
+  const inputRefs = useRef<(HTMLInputElement | null)[]>([]);
+
+  const focusMove = (index: number) => {
+    if (index < totalLength - 1) {
+      inputRefs.current[index + 1]?.focus();
+    }
+  };
+
+  return { inputRefs, focusMove };
+};
